+++ mvc/src/main/java/com/interface21/webmvc/servlet/mvc/tobe/AnnotationHandlerMapping.java
@@ -9,18 +9,22 @@ import java.lang.reflect.Modifier;
 import java.util.HashMap;
 import java.util.Map;
 import java.util.Set;
-import lombok.RequiredArgsConstructor;
 import org.reflections.Reflections;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-@RequiredArgsConstructor
 public class AnnotationHandlerMapping {
 
     private static final Logger log = LoggerFactory.getLogger(AnnotationHandlerMapping.class);
 
     private final Object[] basePackage;
-    private final Map<HandlerKey, HandlerExecution> handlerExecutions = new HashMap<>();
+    private final Map<HandlerKey, HandlerExecution> handlerExecutions;
+
+    public AnnotationHandlerMapping(final Object... basePackage) {
+        this.basePackage = basePackage;
+        this.handlerExecutions = new HashMap<>();
+    }
+
 
     public void initialize() {
         final Reflections reflections = new Reflections(basePackage);
